refactor: group exports in index files
