+++ src/domain/errors/index.ts
@@ -1,0 +1,3 @@
+export * from './invalid-arguments-error'
+export * from './invalid-credentials-error'
+export * from './unexpected-error'
