+++ src/domain/test/mock-authentication.ts
@@ -1,6 +1,6 @@
-import { AuthenticationParams } from '@/domain/usecases/authentication'
 import { faker } from '@faker-js/faker'
-import { AccountModel } from '../models/account-model'
+import { AuthenticationParams } from '@/domain/usecases/authentication'
+import { AccountModel } from '@/domain/models'
 
 export const mockAuthentication = (): AuthenticationParams => ({
   email: faker.internet.email(),
