+++ src/domain/usecases/authentication.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '@/domain/models/account-model'
+import { AccountModel } from '@/domain/models'
 
 export type AuthenticationParams = {
   email: string
